refactor: changed the data type of 'price' from Double to Float
even though the Aggregates on YouTrack says it should be a Double, the offering-service uses Float

+++ src/main/kotlin/eda/shoppingBasket/service/entity/Offering.kt
@@ -8,8 +8,8 @@ import java.util.UUID
 @Entity
 class Offering(
     val name: String,
-    val unitPrice: Double,
-    val totalPrice: Double,
+    val unitPrice: Float,
+    val totalPrice: Float,
     @Id
     val offeringID: UUID,
     @ManyToOne
